requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1  # If you are using SQLAlchemy for database interactions
Flask-Migrate==3.1.0      # If you are using Flask-Migrate for database migrations
pandas==1.5.1              # If you need pandas for data manipulation
# Add any other dependencies your application may require
